+++ src/enum/SupportedFileTypes.ts
@@ -1,0 +1,10 @@
+export enum SupportedFileTypes {
+  HTML = "html",
+  LIQUID = "liquid",
+  JSX = "javascriptreact",
+  VUE = "vue",
+  ANGULAR_TS = "typescript", // Angular TS files may contain ngClass bindings
+  SVELTE = "svelte",
+  PHP = "php",
+  HANDLEBARS = "handlebars",
+}