+++ CHANGELOG.md
@@ -2,8 +2,28 @@
 
 All notable changes to the "classextractor" extension will be documented in this file.
 
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
+## [1.1.1] - 2025-05-10
 
-## [Unreleased]
+### Added
 
-- Initial release
+- Support for additional file types:
+
+  - .vue (Vue.js files)
+  - .ts (Angular TypeScript files)
+  - .svelte (Svelte files)
+  - .php (PHP files)
+  - .handlebars (Handlebar files)
+
+- Extension subscription
+
+## [1.1.1] - 2025-05-10
+
+### Added
+
+- New icon for the extension in the marketplace.
+
+## [1.1.1] - 2025-05-10
+
+### Added
+
+- Initial release with base functionality.